Make meters use a string key, instead of a hash
We were using the lower 64 bits of the md5 of the name and tags as the
map key. It is very unlikely that it will cause collisions with normal
data, but if it did it would be very hard to debug. Replace the hash
with the concatenation of the name and sorted tags as the key.