Use ipc metric names (#40)
Switch from the old spectator names, to the new IPC metric names:

https://netflix.github.io/spectator/en/latest/ext/ipc/

This adds a hook to the config to allow the user to specify which timer
record implementation to use. Usually this is done with a
PercentileTimer but since that is part of the histogram package we
cannot easily use it without creating a circular dependency. The default
is a basic timer.