Fix race in MapKey (#55)
* fix(clock): protect concurrent access to ManualClock fields

The lack of synchronization around accesses to the nanos field trips the
Go race detector in test cases where this type is used.

* fix(http_client): fix incorrect args in test errors

* fix(id): protect access to value for MapKey()

Make MapKey() thread-safe by wrapping the computation with a sync.Once
This does not change any documented behavior of Id, which is immutable
(apart from one case mentioned below).

Interestingly, in the previous implementation, the typical use of
MapKey() was to be first called by the Registry, which uses a mutex for
registration via a blocking call to NewTimer() or friends. I suspect
that the most common use of Id has been to construct and immediately
register. That pattern (and the lack of race detector tests) was able to
hide this bug.

One very slight change in behavior necessitated by the use of sync.Once
is that id.key is now set to the errKey value if an error is encountered
during string building. Previously, a user could try MapKey() again and
attempt another computation of the key. I suspect checking the return
value of MapKey is uncommon, and the error value ("ERR") is not
documented.