fix file descriptor metric names
This commit fixes a bug that was introduced by the following commit:

https://github.com/Netflix/spectator-go/commit/375b638ffabe980c39294486dc6e03a9993a2fb3

When the file descriptor metric names were updated from `fh.maxOpen` and
`fh.curOpen` to `fh.allocated` and `fh.max`, the names were swapped. The
`allocated` metric should track the number of file descriptors opened,
and the `max` metric should track the number of file descriptors
configured.