add payload debugging
There are variety of ways to handle logging in Go, which means that there is
no simple test available for detecting the current log level. As such, this
change introduces an environment variable SPECTATOR_DEBUG_PAYLOAD which can
be set to 1 to enable logging of all Ids and Values in a metrics payload that
is sent to the backend. This value is read once at startup and cached in the
main Registry struct.