log http body on error response when unmarshalling fails
The Atlas response has a field `message` which may be either a string or a
list of strings. The most common case, for validation errors, is a list of
strings. To provide a little more feedback when unusual errors occur, we
will now log the HTTP body for messages that cannot be unmarshalled due to
this type mismatch. This will cover both the single string variant of Atlas
responses, as well as any error responses that may originate from load
balancers.