Add WithCompressThreshold functional arg for NewHttpClient (#69)
* Add WithCompressThreshold functional arg for NewHttpClient

Currently, `HttpClient` hardcodes a 512 byte threshold for enabling
request payload compression. This means the client doesn't work with
servers that don't support `Content-Encoding: gzip`. Adding a knob to
control the threshold makes it possible to disable request compression.

This is technically a breaking API change, since HttpClientOption now
has a different type, but no functional options of this type are exposed
within this source tree, so I'm not sure this was used in the first
place. I've exported `client` inside the `HttpClient` struct, in case
any external code was passing in functions that manipulate the
`net/http` client.

* Add test which sends > 512b of data uncompressed