registry: adds spectator.registrySize meter (#70)
This diff adds the spectator.registrySize meter that is reported by the
Java implementation and spectatord. This dist summary records the total
number of meters in the registry. This is useful if one wants to know
how many meters are in memory and whether that number is significantly
different from the number of measurements being reported.

For the sake of preserving backwards compatability I opted to update
this when `Registry.Measurements` is called as we already unlock the
registry there.

As an alternative, I considered adding a `Registry.Size` method that
would allow callers to report this metric on their own but that creates
more work for callers.