restore Clock interface (#78)
The reason to keep this is that it allows people to easily construct their
own Age Gauges, without relying on the spectatord Age Gauge meter type. The
main reason to do this is so that things like uptime gauges are tied to the
process lifetime. If a process is killed because it exceeded systemd memory
limits, then the spectatord Age Gauge would not be updated.

Example age gauge implementation:

```
func collectUptime(registry *spectator.Registry, uptimeID *meter.Id) {
    bootTime := registry.Clock().Now()
    g := registry.GaugeWithId(uptimeID)
    g.Set(0)
    ticker := time.NewTicker(30 * time.Second)
    go func() {
        for range ticker.C {
            g.Set(registry.Clock().Now().Sub(bootTime).Seconds())
        }
    }()
}
```