remove Clock
In the initial rewrite, the Clock interface was removed, because it was no
longer needed by the internal classes of the thin client. The Clock was added
back to the client, as a means of providing a consistent interface across
spectator clients and to smooth over the differences between the various time
implementations offered by languages.

Upon further reflection, the Clock is not needed, because it is a thin wrapper
around `time.Now()`, with the ability to set specific time values, to support
testing, which is not needed in the thin client. This change removes the Clock
again, and the advice will be to use `time.Now()` directly, as needed, for
timing purposes.

Example age gauge:

```
func collectUptime(registry *spectator.Registry, uptimeID *meter.Id) {
    bootTime := time.Now()
    g := registry.GaugeWithId(uptimeID)
    g.Set(0)
    ticker := time.NewTicker(60 * time.Second)
    go func() {
        for range ticker.C {
            g.Set(time.Now().Sub(bootTime).Seconds())
        }
    }()
}
```