add character validation (#82)
It is easy enough to break spectatord protocol line parsing by using the line
separator characters (,:=) in unexpected ways, such that it may not report a
parsing error. Rather than trying to capture every possible edge case in the
line parsing, the client library can perform basic character validation and
substitution, following the same rules, so that invalid characters are changed
to underscores (_).