switch int64 to uint64 (#90)
In a number of places throughout the code, we were using the `int64` type,
while performing checks to ensure that the values were non-negative. The
better way to do this is to switch the type to `uint64`, which will call
out those errors at compile time, and will help prevent integer overflow
errors, when dealing with very large numbers, as may be the case with
monotonic counters. Switching these types also allows us to delete a fair
amount of code for checking and testing these values.

Of all the type swaps that occurred in this change, only the Counter and
the MonotonicCounter wanted non-zero values. However, these values are not
illegal to send to spectatord, so we are fine dropping that restriction.

Also, update the monotonic counter parameter name to reflect that it is
recording a value, because spectatord is now responsible for calculating
the delta.

Also, run `go mod tidy` to remove all of the unused dependency references.