Revert "switch int64 to uint64 (#90)"
This reverts commit 8d0203bf8d7916a3be7f39d11eb5a1929cf778e5.

There are two issues with this change: (1) it is a breaking API change, and
(2) it is a bit of an un-natural usage pattern, because you may need to cast
variables to `uint64` in order to use these functions. In Go, there are no
literals for `uint64`, so you need to declare that type up-front during the
construction of a variable. If you pass an int literal to the functions, they
will be accepted as long as they fit the parameters of `uint64`.

We should offer support for `uint64`, particularly for `MonotonicCounter`s,
which can grow large, but it should be done through additional methods such
as `AddUint`, much like was done for `float64` in the `AddFloat` functions.
This way, you can use `uint64`, if you need them, but common case of working
with `int` literals will behave as expected.