monotonic counters should use uint64 data type
* There are cases, when working with `MonotonicCounter`s, where it is possible
to experience integer overflow, when using the int64 data type. Using a `uint64`
data type offers a means of working around this issue. Atlas stores all data as
doubles on the backend, and the actual deltas reported to Atlas are expected to
be well within that range. Spectatord provides `MonotonicCounter` as a convenient
way of avoiding the need to track state and calculate deltas, and as such, should
support `uint64`.

* Update the `MonotonicCounter` parameter name to reflect that it is
recording a value, because spectatord is responsible for calculating
the delta.

* Clean up parameter names and docstrings. Sort function names.

* Run `go mod tidy`, to remove all of the unused dependency references.

* Add a Makefile to ease common operations.