add Reset function to MemoryWriter
Since the MemoryWriter was updated to address race conditions, access to the
slice is now private, so it cannot be replaced with an empty slice when you
want to reset it. This is done in the spectator-go-runtime-metrics library,
and may be done in other places.

This change adds a Reset function to the MemoryWriter, which can be used to
reset the storage to an empty slice.