fix monotonic counter implementations
This change is a companion to the fix for monotonic counter implementations
in spectatord.

https://github.com/Netflix-Skunkworks/spectatord/pull/90

The original monotonic counter (`C`) was always intended to be used for the
case where a monotonic data source needs to be transformed into base units
for recording data. For example, transforming nanoseconds into seconds. This
requires division, which results in floats.

There is a valid use case for handling uints in monotonic counters, if the
data source is already in a base unit, such as bytes. Thus, a new meter type
`U` is added to spectatord which supports this use case.