add unix spectatord output location (#105)
Add a new spectatord output location, `unix`, which selects the default Unix
Domain Socket used by spectatord. This saves the user from having to lookup
this value.

Add Registry tests to validate that the with-id methods add any extra common
tags that are specified in the Config. This is true as long as the `NewId`
method on the `Registry` is used to create `Id`s.

In the Config, it is not necessary to check the result from `tagsFromEnvVars`,
because only non-empty values are added to the map.