make UnixgramWriter resilient to connection drops (#108)
We experienced an issue where something was disrupting communication to the
`spectatord` unix socket, which caused the connection to fail every write after
that point. The `UnixgramWriter` was originally configured such that it would
panic if it could not establish an initial connection with the unix socket, so
`spectatord` would start, something would disturb the connection, and then no
more metrics would be delivered and the logs would be spammed with errors.

The two main differences between the `UdpWriter` and the `UnixgramWriter`, are
(1) the `UdpWriter` is more resilient under more conditions, with less logic,
and (2) the unix socket can provide higher throughput.

This change adds resilience to the `UnixgramWriter`, so that initialization
is resilient to the absence of the unix socket, and any disruption of the
connection will cause it to close the connection, and reconnect on subsequent
writes. In order to maintain simplicity and statelessness, some packet loss
will occur during a connection disruption, but it will resume as soon as the
unix socket is available again. As long as metrics are delivered frequently,
more often than once per minute, then small losses such as this may not be
observed.

These changes were tested on a live system, while restarting the `spectatord`
process.

Additionally, clean up handling of the default writer.