enable larger buffer sizes
In highly parallel workloads, the number of metrics sent can be very large,
and we want to reduce contention on the sockets, which are protected by
mutexes. In order to absorb all metrics being sent within the five second
periodic flush window, we may need very large buffers, which still respect
the socket buffer chunk size.